moved Story model to its own file, in StoriesView marked ProgressConfiguration Struct as private
